pyproject.toml
[tool.poetry]
name = "yandex-camp-chat"
version = "0.1.0"
description = ""
authors = ["Ramil Gabdrakhmanov <ramil2911912@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
python-telegram-bot = ">=22.4,<23.0"
openai = ">=1.107.3,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
loguru = ">=0.7.3,<0.8.0"
langchain = ">=0.3.27,<0.4.0"
langchain-openai = ">=0.3.33,<0.4.0"
langchain-community = ">=0.3.29,<0.4.0"
langchain-core = ">=0.3.76,<0.4.0"
boto3 = "^1.40.32"
chromadb = ">=0.5.0,<0.6.0"
sentence-transformers = ">=3.0.0,<4.0.0"
pypdf = ">=4.0.0,<5.0.0"

[tool.poetry.scripts]
bot = "app.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
